Add User crud exceptions

+++ src/modules/user/application/service/user.service.ts
@@ -1,6 +1,6 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, HttpException, HttpStatus } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
-import { Repository } from 'typeorm';
+import { DeleteResult, Repository } from 'typeorm';
 import { User } from '../../infrastructure-persistance/entities/user.entity';
 import { CreateUserDto } from '../../controllers/dto/create-user.dto';
 import { UpdateUserDto } from '../../controllers/dto/update-user.dto';
@@ -10,9 +10,18 @@ export class UserService {
   constructor(
     @InjectRepository(User)
     private userRepository: Repository<User>,
-  ) {}
+  ) { }
 
-  createUser(user: CreateUserDto) {
+  async createUser(user: CreateUserDto) {
+
+    const userFound = await this.userRepository.findOne({
+      where:
+        { username: user.username } || { document: user.document }
+    })
+    if (userFound) {
+      return new HttpException('User already exist', HttpStatus.CONFLICT)
+    }
+
     const newUser = this.userRepository.create(user);
     return this.userRepository.save(newUser);
   }
@@ -21,15 +30,44 @@ export class UserService {
     return this.userRepository.find();
   }
 
-  getUser(id: number) {
-    return this.userRepository.findOne({ where: { id } });
+  async getUser(id: number) {
+    const userFound = await this.userRepository.findOne({ where: { id } });
+
+    if (!userFound) {
+      return new HttpException('User not found', HttpStatus.NOT_FOUND);
+    }
+
+    return userFound;
   }
 
-  deleteUser(id: number){
-    return this.userRepository.delete({ id });
+  async deleteUser(id: number): Promise<DeleteResult | HttpException> {
+    const userFound = await this.userRepository.findOne({ where: { id }});
+  
+    if (!userFound){
+      return new HttpException('User not found', HttpStatus.NOT_FOUND);
+    }
+  
+    return this.userRepository.delete( { id });
+
+     /* 
+    const result = await this.userRepository.delete({ id });
+
+    if (result.affected === 0){
+      return new HttpException('User not found', HttpStatus.NOT_FOUND);
+    }
+
+    return result;
+    */
   }
 
-  updateUser(id: number, user: UpdateUserDto){
-    return this.userRepository.update({ id }, user);
+  async updateUser(id: number, user: UpdateUserDto): Promise<User | HttpException> {
+    const userFound = await this.userRepository.findOne({where: { id }});
+
+    if (!userFound){
+      return new HttpException('User not found', HttpStatus.NOT_FOUND);
+    }
+
+    const updatedUser = Object.assign(userFound, user);
+    return this.userRepository.save(updatedUser);
   }
 }
